+++ src/main/java/chess/domain/controller/dto/PieceResponse.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.domain.controller.dto;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
